combine all operation
