--- backend/er model+shema/test.png
@@ -1,1 +1,0 @@
-
