--- backend/frontend/templates/delete.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Bike store operation site</title>
-    <link rel="stylesheet" type="text/css" href="../static/css/style.css" />
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-    <script src="../static/js/scripts.js" defer></script>
-  </head>
-
-  <body>
-    <div class="topbar">
-      <!--<img src="" width="80px" height="80px" alt="Log">-->
-      <p> Bike store operation site </p>
-      <div class="menu">
-        <a href="/">Home</a>
-        <a href="/dashboard">Dashboard</a>
-        <a href="/new">新增訂單</a>
-        <a href="/delete">刪除訂單</a>
-        <a href="/search">搜尋訂單</a>
-        <a href="/update">更新訂單</a>
-        <a href="/update_order_item">更新訂單項目</a>
-      </div>
-
-    </div>
-    <div class="title flex-position">
-      <p>Delete Order</p>
-    </div>
-    <form action="/delete_tackle" method="post">
-      <div>
-        <div class="container">
-          <input type="text" id="order_id" name="order_id" placeholder="請輸入您欲刪除的order_id "/>
-        </div>
-        <div class="container">
-          <button>Submit</button>
-        </div>
-      </div>
-    </form>
-    <div class="short-thick-line"></div>
-    <div class="bottom-text">
-      <p>© 2024 Abner Den</p>
-    </div>
-  </body>
-</html>
