--- backend/frontend/templates/index.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Bike store operation site</title>
-    <link rel="stylesheet" type="text/css" href="../static/css/style.css">
-</head>
-
-<body>
-    <div class="topbar">
-        <img src="" width="80px" height="80px" alt="Log">
-        <p> Bike store operation site </p>
-        <div class="menu">
-            <a href="/">Home</a>
-            <a href="/dashboard">Dashboard</a>
-            <a href="/new">新增訂單</a>
-            <a href="/delete">刪除訂單</a>
-            <a href="/search">搜尋訂單</a>
-            <a href="/update">更新訂單</a>
-            <a href="/update_order_item">更新訂單項目</a>
-
-        </div>
-        <div class="logreg-container">
-            <div>
-                <button class="log" onclick="window.location.href='login.html'">Login</button>
-                <button class="reg" onclick="window.location.href='signup.html'">Sign up</button>
-            </div>
-        </div>
-    </div>
-    <div class="flex-main">
-        <div class="main-text">
-            <h1>腳踏車銷售系統</h1>
-        </div>
-        <div class="container">
-            <div class="content">
-                <h2>請點選右上角的功能進行 CRUD 四個功能!</h2>
-            </div>
-        </div>
-    </div>
-    <div class="short-thick-line"></div>
-    <div class="bottom-text">
-        <p>© 2024 Abner Den</p>
-    </div>
-</body>
-
-</html>
