--- backend/frontend/templates/new.html
@@ -1,90 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Bike store operation site</title>
-    <link rel="stylesheet" type="text/css" href="../static/css/style.css" />
-    <script src="../static/js/scripts.js" defer></script>
-  </head>
-
-  <body>
-    <div class="topbar">
-      <!-- <img src="" width="80px" height="80px" alt="Log" />-->
-      <p>Bike Store operation site</p>
-      <div class="menu">
-        <a href="/">Home</a>
-        <a href="/dashboard">Dashboard</a>
-        <a href="/new">新增訂單</a>
-        <a href="/delete">刪除訂單</a>
-        <a href="/search">搜尋訂單</a>
-        <a href="/update">更新訂單</a>
-        <a href="/update_order_item">更新訂單項目</a>
-      </div>
-      <div class="logreg-container">
-        <div>
-          <button class="log">Login</button>
-          <button class="reg">Sign up</button>
-        </div>
-      </div>
-    </div>
-    <div class="flex-position">
-      <div class="title">
-        <p>Create Order</p>
-      </div>
-    </div>
-    <form action="/new_tackle" method="post">
-      <div class="flex-position">
-          <div class="form-group">
-              <label for="customer_id">customer_id :</label>
-              <input type="text" id="customer_id" name="customer_id" />
-          </div>
-          <div class="form-group">
-              <label for="order_status">order_status :</label>
-              <input type="text" id="order_status" name="order_status" />
-          </div>
-          <div class="form-group">
-              <label for="order_date">order_date :</label>
-              <input type="date" id="order_date" name="order_date" />
-          </div>
-          <div class="form-group">
-              <label for="required_date">required_date :</label>
-              <input type="date" id="required_date" name="required_date" />
-          </div>
-          <div class="form-group">
-              <label for="shipped_date">shipped_date :</label>
-              <input type="date" id="shipped_date" name="shipped_date" />
-          </div>
-          <div class="form-group">
-              <label for="store_id">store_id :</label>
-              <input type="text" id="store_id" name="store_id" />
-          </div>
-          <div class="form-group">
-              <label for="staff_id">staff_id :</label>
-              <input type="text" id="staff_id" name="staff_id" />
-          </div>
-          <div class="form-group">
-              <label for="product_id">product_id :</label>
-              <input type="text" id="product_id" name="product_id" placeholder="若有多項產品請以,分隔"/>
-          </div>
-          <div class="form-group">
-              <label for="quantity">quantity :</label>
-              <input type="text" id="quantity" name="quantity" placeholder="若有多項產品請以,分隔"/>
-          </div>
-          <div class="form-group">
-              <label for="discount">discount :</label>
-              <input type="text" id="discount" name="discount" />
-          </div>
-
-      </div>
-      <div class="container">
-        <button type="submit">Submit Order</button>
-    </div>
-  </form>
-    <div class="short-thick-line"></div>
-    <div class="bottom-text">
-      <p>© 2024 Abner Den</p>
-    </div>
-  </body>
-</html>
-
