--- backend/frontend/templates/result.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Update Result</title>
-</head>
-<body>
-    <h1>{{ result }}</h1>
-    <a href="/">返回主頁!!</a>
-</body>
-</html>
