--- backend/frontend/templates/signup.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Bike store operation site</title>
-    <link rel="stylesheet" type="text/css" href="../static/css/style.css">
-    <link rel="stylesheet" type="text/css" href="../static/css/signup.css">
-    <script src="../static/js/signup.js" defer></script>
-</head>
-
-<body>
-    <h1>Sign Up</h1>
-    <form id="signup-form">
-        <label for="username">Username</label>
-        <input type="text" id="username" name="username" required><br>
-        <label for="password">Password</label>
-        <input type="password" id="password" name="password" required><br>
-        <button type="submit">Sign Up</button>
-    </form>
-</body>
-
-</html>
