--- backend/frontend/templates/update.html
@@ -1,78 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Bike store operation site</title>
-    <link rel="stylesheet" type="text/css" href="../static/css/style.css" />
-    <script src="../static/js/scripts.js" defer></script>
-  </head>
-
-  <body>
-    <div class="topbar">
-      <img src="" width="80px" height="80px" alt="Log" />
-      <p>Bike Store operation site</p>
-      <div class="menu">
-        <a href="/">Home</a>
-        <a href="/dashboard">Dashboard</a>
-        <a href="/new">新增訂單</a>
-        <a href="/delete">刪除訂單</a>
-        <a href="/search">搜尋訂單</a>
-        <a href="/update">更新訂單</a>
-        <a href="/update_order_item">更新訂單項目</a>
-      </div>
-      <div class="logreg-container">
-        <div>
-          <button class="log">Login</button>
-          <button class="reg">Sign up</button>
-        </div>
-      </div>
-    </div>+
-    <div class="flex-position">
-      <div class="title">
-        <p>Update Order</p>
-      </div>
-    </div>
-    <form action="/update_tackle" method="post">
-      <div>
-        <label for="order_id">請輸入您欲更新的order_id :</label>
-        <input type="text" id="order_id" name="order_id" required />
-      </div>
-      <div>
-        <label for="customer_id">顧客ID:</label>
-        <input type="text" id="customer_id" name="customer_id" required />
-      </div>
-      <div>
-        <label for="order_status">訂單狀態:</label>
-        <input type="text" id="order_status" name="order_status" required />
-      </div>
-      <div>
-        <label for="order_date">訂單日期:</label>
-        <input type="date" id="order_date" name="order_date" required />
-      </div>
-      <div>
-        <label for="required_date">需求日期:</label>
-        <input type="date" id="required_date" name="required_date" required />
-      </div>
-      <div>
-        <label for="shipped_date">出貨日期:</label>
-        <input type="date" id="shipped_date" name="shipped_date" />
-      </div>
-      <div>
-        <label for="store_id">店鋪ID:</label>
-        <input type="text" id="store_id" name="store_id" required />
-      </div>
-      <div>
-        <label for="staff_id">員工ID:</label>
-        <input type="text" id="staff_id" name="staff_id" required />
-      </div>
-      <div>
-        <button type="submit">Submit</button>
-      </div>
-    </form>
-    <div class="short-thick-line"></div>
-    <div class="bottom-text">
-      <p>© 2024 Abner Den</p>
-    </div>
-  </body>
-</html>
