--- backend/frontend/templates/update_order_item.html
@@ -1,69 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Update Order Item</title>
-    <link rel="stylesheet" type="text/css" href="../static/css/style.css">
-    <script src="../static/js/scripts.js" defer></script>
-</head>
-<body>
-    <div class="topbar">
-        <img src="" width="80px" height="80px" alt="Logo">
-        <p>Bike Store Operation Site</p>
-        <div class="menu">
-            <a href="/">Home</a>
-            <a href="/dashboard">Dashboard</a>
-            <a href="/new">新增訂單</a>
-            <a href="/delete">刪除訂單</a>
-            <a href="/search">搜尋訂單</a>
-            <a href="/update">更新訂單</a>
-            <a href="/update_order_item">更新訂單項目</a>
-        </div>
-        <div class="logreg-container">
-            <div>
-                <button class="log">Login</button>
-                <button class="reg">Sign up</button>
-            </div>
-        </div>
-    </div>
-    <div class="flex-position">
-        <div class="title">
-            <p>Update Order Item</p>
-        </div>
-    </div>
-    <form action="/update_order_item_tackle" method="post">
-        <div>
-            <label for="order_id">請輸入您欲更新的 order_id :</label>
-            <input type="text" id="order_id" name="order_id" required>
-        </div>
-        <div>
-            <label for="item_id">請輸入您欲更新的 item_id :</label>
-            <input type="text" id="item_id" name="item_id" required>
-        </div>
-        <div>
-            <label for="product_id">產品ID:</label>
-            <input type="text" id="product_id" name="product_id" required>
-        </div>
-        <div>
-            <label for="quantity">數量:</label>
-            <input type="number" id="quantity" name="quantity" required>
-        </div>
-        <div>
-            <label for="list_price">價格:</label>
-            <input type="number" step="0.01" id="list_price" name="list_price" required>
-        </div>
-        <div>
-            <label for="discount">折扣:</label>
-            <input type="number" step="0.01" id="discount" name="discount" required>
-        </div>
-        <div>
-            <button type="submit">Submit</button>
-        </div>
-    </form>
-    <div class="short-thick-line"></div>
-    <div class="bottom-text">
-        <p>© 2024 Abner Den</p>
-    </div>
-</body>
-</html>
