+++ frontend/templates/dashboard.html
@@ -18,12 +18,12 @@
         <p> Bike store operation site </p>
         <div class="menu">
             <a href="/">Home</a>
-            <a href="/crud">CRUD</a>
-            <a href="/dashboard">Dashboard</a> 
-            <!-- <a href="index.html">Home</a>
-            <a href="crud.html">CRUD</a>
-            <a href="dashboard.html">Dashboard</a>
-            -->
+            <a href="/dashboard">Dashboard</a>
+            <a href="/new">新增訂單</a>
+            <a href="/delete">刪除訂單</a>
+            <a href="/search">搜尋訂單</a>
+            <a href="/update">更新訂單</a>
+            <a href="/update_order_item">更新訂單項目</a>
         </div>
         <div class="logreg-container">
             <div>
