+++ frontend/templates/delete.html
@@ -14,27 +14,15 @@
       <!--<img src="" width="80px" height="80px" alt="Log">-->
       <p> Bike store operation site </p>
       <div class="menu">
-          <!-- path for flask api
-              <a href="/">Home</a>
-              <a href="/service1">service1</a>
-              <a href="/service2">service2</a>
-              <a href="/new">新增訂單</a>
-              <a href="/delete">刪除訂單</a>
-              <a href="/update">更新訂單</a>
-              <a href="/update_order_item">更新訂單項目</a> 新增更新訂單項目的鏈接
-          -->
-          <a href="index.html">Home</a>
-          <a href="dashboard.html">Dashboard</a>
-          <a href="new.html">Create Order</a>
-          <a href="update.html">Update Order</a>
-          <a href="update_order_item.html">Update Order Item</a>
+        <a href="/">Home</a>
+        <a href="/dashboard">Dashboard</a>
+        <a href="/new">新增訂單</a>
+        <a href="/delete">刪除訂單</a>
+        <a href="/search">搜尋訂單</a>
+        <a href="/update">更新訂單</a>
+        <a href="/update_order_item">更新訂單項目</a>
       </div>
-      <div class="logreg-container">
-        <div>
-            <button class="log">Login</button>
-            <button class="reg">Sign up</button>
-        </div>
-      </div>
+
     </div>
     <div class="title flex-position">
       <p>Delete Order</p>
