+++ frontend/templates/index.html
@@ -14,20 +14,13 @@
         <p> Bike store operation site </p>
         <div class="menu">
             <a href="/">Home</a>
-            <a href="/crud">CRUD</a>
             <a href="/dashboard">Dashboard</a>
             <a href="/new">新增訂單</a>
             <a href="/delete">刪除訂單</a>
+            <a href="/search">搜尋訂單</a>
             <a href="/update">更新訂單</a>
-            <a href="/update_order_item">更新訂單項目</a> 新增更新訂單項目的鏈接
+            <a href="/update_order_item">更新訂單項目</a>
 
-            <!-- <a href="index.html">Home</a>
-            <a href="crud.html">CRUD</a>
-            <a href="dashboard.html">Dashboard</a>
-            <a href="new.html">Create Order</a>
-            <a href="delete.html">Delete Order</a>
-            <a href="update.html">Update Order</a>
-            <a href="update_order_item.html">Update Order Item</a> -->
         </div>
         <div class="logreg-container">
             <div>
@@ -38,20 +31,11 @@
     </div>
     <div class="flex-main">
         <div class="main-text">
-            <h1>temp</h1>
+            <h1>腳踏車銷售系統</h1>
         </div>
         <div class="container">
             <div class="content">
-                <h2>temp</h2>
-                <li>item1</li>
-                <li>item2</li>
-            </div>
-            <div class="vertical-line"></div>
-            <div class="content">
-                <h2>temp</h2>
-                <li>item1</li>
-                <li>item2</li>
-                <li>item3</li>
+                <h2>請點選右上角的功能進行 CRUD 四個功能!</h2>
             </div>
         </div>
     </div>
