+++ frontend/templates/result.html
@@ -1,15 +1,11 @@
 <!DOCTYPE html>
 <html lang="en">
-
 <head>
     <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Image Processing Results</title>
+    <title>Update Result</title>
 </head>
-
 <body>
-    <h1>Image Processing Complete!</h1>
-    <img src="{{ url_for('static', filename=output) }}" alt="Result Image">
+    <h1>{{ result }}</h1>
+    <a href="/">返回主頁!!</a>
 </body>
-
 </html>
