+++ frontend/templates/search_manager.html
