+++ frontend/templates/update.html
@@ -13,14 +13,13 @@
       <img src="" width="80px" height="80px" alt="Log" />
       <p>Bike Store operation site</p>
       <div class="menu">
-        <!-- path for flask api
-                <a href="/">Home</a>
-                <a href="/service1">service1</a>
-                <a href="/service2">Service2</a> 
-            -->
-        <a href="index.html">Home</a>
-        <a href="service1.html">Service1</a>
-        <a href="service2.html">Service2</a>
+        <a href="/">Home</a>
+        <a href="/dashboard">Dashboard</a>
+        <a href="/new">新增訂單</a>
+        <a href="/delete">刪除訂單</a>
+        <a href="/search">搜尋訂單</a>
+        <a href="/update">更新訂單</a>
+        <a href="/update_order_item">更新訂單項目</a>
       </div>
       <div class="logreg-container">
         <div>
@@ -28,7 +27,7 @@
           <button class="reg">Sign up</button>
         </div>
       </div>
-    </div>
+    </div>+
     <div class="flex-position">
       <div class="title">
         <p>Update Order</p>
