+++ frontend/templates/update_order_item.html
@@ -12,14 +12,13 @@
         <img src="" width="80px" height="80px" alt="Logo">
         <p>Bike Store Operation Site</p>
         <div class="menu">
-            <!-- path for flask api
             <a href="/">Home</a>
-            <a href="/service1">service1</a>
-            <a href="/service2">Service2</a> 
-            -->
-            <a href="index.html">Home</a>
-            <a href="service1.html">Service1</a>
-            <a href="service2.html">Service2</a>
+            <a href="/dashboard">Dashboard</a>
+            <a href="/new">新增訂單</a>
+            <a href="/delete">刪除訂單</a>
+            <a href="/search">搜尋訂單</a>
+            <a href="/update">更新訂單</a>
+            <a href="/update_order_item">更新訂單項目</a>
         </div>
         <div class="logreg-container">
             <div>
